server/requirements.txt
flask==3.0.3
flask-cors==5.0.0
werkzeug==3.0.3
langchain==0.2.16
langchain-chroma==0.1.4
langchain-community==0.2.16
langchain-core==0.2.39
langchain-experimental==0.0.65
langchain-ollama==0.1.3
chromadb==0.4.24
pydantic>=2.7.4
pypdf
python-dotenv